move the parsing to new class Parser
